requirements_cpu.txt
faiss-cpu==1.7.4
numpy==1.24.4
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
altair==5.4.0
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.1.0
cryptography==43.0.3
cycler==0.12.1
datasets==3.1.0
dill==0.3.8
easyocr==1.7.2
exceptiongroup==1.2.2
fastapi==0.115.11
ffmpy==0.5.0
filelock==3.16.1
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.9.0
gradio==3.40.1
gradio_client==1.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.23.0
idna==3.10
imageio==2.35.1
importlib_resources==6.4.5
intel-cmplr-lib-ur==2025.0.5
intel-openmp==2025.0.5
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
lazy_loader==0.4
linkify-it-py==2.0.3
lxml==5.3.1
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdit-py-plugins==0.3.3
mdurl==0.1.2
mkl==2025.0.1
mkl-service==2.4.1
modelscope==1.23.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.31.0
networkx==3.1
ninja==1.11.1.3
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
orjson==3.10.15
packaging==24.2
pandas==2.0.3  
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==10.4.0
propcache==0.2.0
pyarrow==17.0.0
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
PyMuPDF==1.24.1
pyparsing==3.1.4
pypdfium2==4.30.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rpds-py==0.20.1
safetensors==0.5.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
semantic-version==2.10.0
shapely==2.0.7
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.44.0
sympy==1.13.1
tbb==2022.0.0
tcmlib==1.2.0
threadpoolctl==3.5.0
tifffile==2023.07.10
tokenizers==0.15.2
--index-url https://download.pytorch.org/whl/cpu
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
--index-url https://pypi.org/simple
tqdm==4.67.1
transformers==4.37.0
typing_extensions==4.12.2
tzdata==2025.1
uc-micro-py==1.0.3
umf==0.9.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==11.0.3
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
accelerate==1.6.0